Yugabyte: multi-key-acid test review
This test used identical values for both keys during writes, rather than
arbitrary values, and created a very large state space. It also only
tested 100 operations *ever*: most of the time it did nothing. We've
moved to a more compact but randomized state space, and updated the test
to use `jepsen.independent`, so it continues testing over the course of
an entire test.